+++ examples/java/core/schema_view/src/main/java/com/microsoft/bond/examples/SchemaView.java
@@ -24,14 +24,14 @@ public class SchemaView {
         example.items.add(0D);
 
         final ByteArrayOutputStream output = new ByteArrayOutputStream();
-        final CompactBinaryWriter writer = new CompactBinaryWriter(output, (short) 1);
+        final CompactBinaryWriter writer = new CompactBinaryWriter(output, 1);
 
         final Serializer<Example> serializer = new Serializer<>();
         serializer.serialize(example, writer);
 
         final ByteArrayInputStream input = new ByteArrayInputStream(output.toByteArray());
 
-        final CompactBinaryReader reader = new CompactBinaryReader(input, (short) 1);
+        final CompactBinaryReader reader = new CompactBinaryReader(input, 1);
         final Deserializer<ExampleView> deserializer = new Deserializer<>(ExampleView.BOND_TYPE);
         final ExampleView view = deserializer.deserialize(reader);
