+++ java/core/src/test/java/com/microsoft/bond/protocol/CompactBinaryV1ProtocolTest.java
@@ -12,7 +12,7 @@ import java.io.IOException;
 
 public class CompactBinaryV1ProtocolTest {
 
-    private static final short PROTOCOL_VERSION = 1;
+    private static final int PROTOCOL_VERSION = 1;
 
     private static final TaggedBinaryTestHelper.ProtocolImplementationFactory FACTORY =
             new TaggedBinaryTestHelper.ProtocolImplementationFactory(
