+++ java/core/src/test/java/com/microsoft/bond/protocol/FastBinaryReaderTest.java
@@ -19,25 +19,25 @@ public class FastBinaryReaderTest {
     @Test(expected = IllegalArgumentException.class)
     public void testConstructorWithZeroProtocolVersion() {
         ByteArrayInputStream bais = new ByteArrayInputStream(new byte[0]);
-        new FastBinaryReader(bais, (short) 0);
+        new FastBinaryReader(bais, 0);
     }
 
     @Test(expected = IllegalArgumentException.class)
     public void testConstructorWithInvalidProtocolVersion() {
         ByteArrayInputStream bais = new ByteArrayInputStream(new byte[0]);
-        new FastBinaryReader(bais, (short) 2);
+        new FastBinaryReader(bais, 2);
     }
 
     @Test(expected = IllegalArgumentException.class)
     public void testConstructorWithNullInputStream() {
         ByteArrayInputStream bais = null;
-        new FastBinaryReader(bais, (short) 1);
+        new FastBinaryReader(bais, 1);
     }
 
     @Test
     public void testSkipWithInvalidDataType() {
         ByteArrayInputStream bais = new ByteArrayInputStream(new byte[0]);
-        FastBinaryReader reader = new FastBinaryReader(bais, (short) 1);
+        FastBinaryReader reader = new FastBinaryReader(bais, 1);
         BondDataType dataType = BondDataType.BT_UNAVAILABLE;
         try {
             reader.skip(dataType);
